Add: CheckRoomId 작성, loader 제거하고 CheckRoomId 컴포넌트로 roomId 확인 대체
